243: move BlockSpec and BlockTag
