+++ src/rpc_client/eth_rpc/mod.rs
@@ -6,7 +6,7 @@ use crate::log;
 use crate::memory::next_request_id;
 use crate::providers::resolve_rpc_service;
 use crate::rpc_client::eth_rpc_error::{sanitize_send_raw_transaction_result, Parser};
-use crate::rpc_client::numeric::{BlockNumber, TransactionCount, Wei};
+use crate::rpc_client::numeric::{TransactionCount, Wei};
 use crate::rpc_client::responses::{Block, FeeHistory, LogEntry, TransactionReceipt};
 use crate::types::MetricRpcMethod;
 use candid::candid_method;
@@ -22,6 +22,7 @@ use minicbor::{Decode, Encode};
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use std::fmt;
 use std::fmt::{Debug, Display, Formatter, LowerHex, UpperHex};
+use crate::rpc_client::requests::BlockSpec;
 
 #[cfg(test)]
 mod tests;
@@ -142,76 +143,6 @@ impl std::str::FromStr for Hash {
 
 impl HttpResponsePayload for Hash {}
 
-/// Block tags.
-/// See <https://ethereum.org/en/developers/docs/apis/json-rpc/#default-block>
-#[derive(Debug, Default, Copy, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub enum BlockTag {
-    /// The latest mined block.
-    #[default]
-    #[serde(rename = "latest")]
-    Latest,
-    /// The latest safe head block.
-    /// See
-    /// <https://www.alchemy.com/overviews/ethereum-commitment-levels#what-are-ethereum-commitment-levels>
-    #[serde(rename = "safe")]
-    Safe,
-    /// The latest finalized block.
-    /// See
-    /// <https://www.alchemy.com/overviews/ethereum-commitment-levels#what-are-ethereum-commitment-levels>
-    #[serde(rename = "finalized")]
-    Finalized,
-    #[serde(rename = "earliest")]
-    Earliest,
-    #[serde(rename = "pending")]
-    Pending,
-}
-
-impl Display for BlockTag {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Latest => write!(f, "latest"),
-            Self::Safe => write!(f, "safe"),
-            Self::Finalized => write!(f, "finalized"),
-            Self::Earliest => write!(f, "earliest"),
-            Self::Pending => write!(f, "pending"),
-        }
-    }
-}
-
-/// The block specification indicating which block to query.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(untagged)]
-pub enum BlockSpec {
-    /// Query the block with the specified index.
-    Number(BlockNumber),
-    /// Query the block with the specified tag.
-    Tag(BlockTag),
-}
-
-impl Default for BlockSpec {
-    fn default() -> Self {
-        Self::Tag(BlockTag::default())
-    }
-}
-
-impl std::str::FromStr for BlockSpec {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if s.starts_with("0x") {
-            let block_number = BlockNumber::from_str_hex(s)
-                .map_err(|e| format!("failed to parse block number '{s}': {e}"))?;
-            return Ok(BlockSpec::Number(block_number));
-        }
-        Ok(BlockSpec::Tag(match s {
-            "latest" => BlockTag::Latest,
-            "safe" => BlockTag::Safe,
-            "finalized" => BlockTag::Finalized,
-            _ => return Err(format!("unknown block tag '{s}'")),
-        }))
-    }
-}
-
 /// Parameters of the [`eth_getBlockByNumber`](https://ethereum.org/en/developers/docs/apis/json-rpc/#eth_getblockbynumber) call.
 #[derive(Debug, Serialize, Clone)]
 #[serde(into = "(BlockSpec, bool)")]
@@ -229,12 +160,6 @@ impl From<GetBlockByNumberParams> for (BlockSpec, bool) {
 }
 
 impl HttpResponsePayload for Wei {}
-
-impl From<BlockNumber> for BlockSpec {
-    fn from(value: BlockNumber) -> Self {
-        BlockSpec::Number(value)
-    }
-}
 
 /// An envelope for all JSON-RPC requests.
 #[derive(Clone, Serialize, Deserialize)]
